🚚 rename: 메인케러셀 인포 파일명 변경 #264
